\docType{data}
    \name{LM22}
    \alias{LM22}
    \title{Signature matrix including expression of 22 immune cell types}
    \format{Gene by cell-type matrix containing the expression profiles of 22 immune cells.}
    \source{
      Cibersort
    }
    \description{
     This data contains expression profiles of 22 immune cells.
    }
    \author{
Newman, Aaron M., Chih Long Liu, Michael R. Green, Andrew J. Gentles, Weiguo Feng, Yue Xu, Chuong D. Hoang, Maximilian Diehn, and Ash A. Alizadeh. "Robust enumeration of cell subsets from tissue expression profiles." Nature methods 12, no. 5 (2015): 453-457.
    }
    \usage{CpG.human.GRCh37}
    \keyword{datasets}
    
